Fix mysqueezebox.com disconnects
At random times (reported between 11 hours and a couple of days), when
connected to mysqueezebox.com a serv command is issued to connect to
server 0.0.0.1:3483. However, since that server is unlikely to exist,
Squeezelite is send in a loop attempting to connect to that new server.
This fix gives it one go, if the new server doesn’t exist, it will
revert to the previous server.